Optionally generate FIT image
SquashFS is great. But not all bootloaders are great.

Sometimes you just want to do a quick test of NetBox on a system with
a less than great bootloader, e.g. most eval boards.

This new option takes the regular SquashFS image and automagically
generates an FIT image from it so that you can boot it from U-Boot.