fitimage.sh: Generate an FIT with predictable configuration names
The Netbox naming convention under /boot means that we have a name for
each supported product. Leverage this information when generating the
FIT such that each configuration is named after the device-tree that
it will use. This means that U-Boot can run something like:

bootm $LOAD#0000-my-board

Instead of:

bootm $LOAD#conf-N

Where N could change whenever a new board is added.

As a bonus we also get to name the images properly.