os: Automatically probe/setup DSA channel interfaces
Instead of relying on /etc/mactab to rename channel interfaces, we can
probe for them based on the existence of /sys/class/net/*/dsa/tagging.

This means you can have two boards, using for example the FEC ethernet
controller, where one board has a switch attached to it and the other
one does not. In the first case it would be renamed to "chan0", in the
other it would be left untouched.