os: split NetBox meta package into three "profiles"
A NetBox build is one of Application or Operating System.  The former is
basically just a plain root filesystem packaged as a squashfs image.
The latter is a root filesystem and a kernel.

At Westermo the kernel image is inside the squashfs image, in the /boot
directory like a normal system.  Some bootloaders do not (yet) support
this so the images directory also saves the kernel image separately.

Each of the "app" and "os" types can then be given a pre-select set of
daemons and tools, called Profiles.  For apps there is a pre-select
Application profile with only Fint + Dropbear SSH + tcpdump, and for
os builds there is an even greater list.  There is no restriction here
for how to compbine a an app type with a specific profile, and for full
freedom there is a custom profile which only selects Finit.

This patch splits the NetBox meta package into profiles and also updates
all the deconfigs.  The app defconfigs have been given the app profile,
so as of this change, a lot less tools/daemons are included for apps by
default.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>