os: initial support for GNS3 appliances with a bootable .iso image
This patch adds initial support for creating GNS3 appliances to the
NetBox Zero OS builds.  We set up GNS3 to start with -cdrom iso and
an empty Ext3 Config partition for storage.  As long as the md5sum
for config.ext3 remains the same, the user should be able to do an
appliance upgrade without losing settings.

Signed-off-by: Joachim Wiberg <joachim.wiberg@westermo.com>