net: dsa: Fix cmode handle on the Marvell Peridot switch
This patch fixes of the cmode "hang-up" problem on the Marvell Peridot switch.
Problem is that You can not change the mac cmode setting in the Peridot in
between a serdes power off and on. Needs to be done in the right order
phy/serdes/mac.

As a result of this fix, You will be able to change the mac cmode and get i.e
the SFPs on the envoy platform to work.

Signed-off-by: Joacim Zetterling <joacim.zetterling@westermo.com>