.github: cache keys seemingly must be uniqe per job
Maybe the downloads can be shared between jobs, but I don't think it's
worth the added complexity.  Because, problems arise in the post cache
step, if a cache is shared only one can write to it ... so then we need
to add dependencys etc.  So let's keep it simple and rely on the free
disk space we have anyway.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>