Drop -dev version from developer builds, use "" as GNS3 version
After a long discussion, this turned out to be the least contentious
alternative to: -dev, -devel, -HEAD, etc.  We now revert back to the
original netbox-os-zero.img, etc., for builds generated by cloning
and building the tree as-is, with modifications or not.

Actual versioned images should only be generated from tagged releases.

Information about the current image is always available in the files
/etc/version and /etc/os-release. There it is also possible to track
actual gHASH and if the build was dirty or not.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>