os: centralize image naming and fix md5sum generation
Resulting image names must be kept beside each other to ensure they all
follow the same naming scheme.  Instead of decentralizing we now send
the image names as function arguments.

We have platforms with multiple images, .iso, .img, .itb, etc.  that we
need to differentiate between, so the .md5 extension should therefore be
added to the end of that extension, not replace it.

Also, the contents of the .md5 should not contain the path to the file.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>