.github: build: revert to compressed tarball
Billing per upload bytes, https://github.com/actions/upload-artifact, so
if we ever need to go private, keep uploads as small as possible.

The main reason for this channge, however, is to retain file modes and a
containing directory for all the files.  Downloading an artifact always
creates a zip of uploaded artifacts, but since it's impossible to keep a
containing directory it forces us to have a zip of a tar.gz ... which is
not ideal, but something we can live with.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>