Simplified README.md to make it easier for beginners to get started.
Added python as a required build dependency.

In order to utilize `make run` as of now, one must have built one of the
os images. Added an example to highlight this.

One also have to add `/etc/qemu/bridge.conf`. On Ubuntu 20.04, this
file does not exist by default (in fact, the directory `/etc/qemu`
does not even exist!).

Signed-off-by: Jörgen Sigvardsson <jorgen.sigvardsson@gmail.com>