package/skeleton-init-finit: add /etc/default support to sysklogd
This patch moves the command line options for syslogd (sysklogd) from
the Finit config to /etc/default/sysklogd.  All execept -F, which is
needed by Finit to run syslogd in the foreground.

Also, rename service from syslogd -> sysklogd to prevent confusion with
the BusysBox syslogd service, which is disabled by default in NetBox.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>