board/common: add 'factory' user with /sbin/factory as login shell
This patch adds support for multiple users to NetBox.  The first user,
created for us by the Buildroot mkusers script, is 'factory' with the
password 'reset'.

All users in the wheel group are allowed to:

 - Create files in /mnt
 - Call the initctl program, e.g. to reboot or stop/restart services

When logging in to the system from the system console, /sbin/factory
is started.  This program touches a file in /mnt and then calls the
reboot program (initctl reboot).

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>