Replace factory reset script with proper C program
Scripts are great for many purposes, but you can't make the SUID root.
This patch imports a new (local) C based tool with the same name that
does the same thing, which we can set to SUID root using xattrs.

When used as login "shell", the factory user is in the wheel group, so
this patch also changes the group and permission modes of the initctl
socket, /run/finit/socket, by backporting parts of an upcoming patch to
Finit.  This also allows other users, part the wheel group, to use the
initctl tool to start/stop services and reboot the system.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>