board/common: update 'show' command with tips, extend 'show ifaces'
For 'show ifaces', skip 'eth.*' interfaces since they are probably part
of the bridge.  This is not ideal of course, a better filter would be to
query the bridge for each interface (bridge port member or not?), but it
was quite hard doing all this from awk so I ended up with this version.

It's better of course that people learn the real Linux commands, hence
the new tips at the bottom of each output.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>