configs: remap std m4 quoutes to autoconf style
This is a usability improvement to the m4 processor used in NetBox to
create the Buildroot defconfig files.  By default the m4 language uses
`' for quoting, not only is this tedious and error prone to write, but
it also does not invite editors to do matching of start<->end, so it
becomes a hassle to debug as well.

After much deliberation we decided to go with the GNU Autoconf [] style
for quouting instead.  This allows for embedding shell strings and doing
back tick execution, and more, from the include files.  And to top
things off, your editor is also happy.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>