configs: simplify, drop dnl at EOL, use dnl in snippet comments
For "perfect" pasting of snippets, we used 'dnl' at end of line.  This
tells m4 to not inject a newline (empty line) in the output.  However,
this can be confusing to newcomers, and is not really necessary since
we feed the resulting defconfig to Buildroot to generate the .config,
and then we remove the defconfig ... so to keep things simple we just
drop this.

This patch also replaces shell comments with dnl, for consistency and
point out that the comments are for maintainers of the defconfigs.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>