board: tell logger to log to /dev/kmsg if syslogd is not up yet
It's a bit messy to output to the console at the same time as the Finit
progress is ouput.  It's also very hard to see when you don't have the
console connected, or it otherwise scrolls out of view.

This patch allows these helper scripts to log to the kernel ring buffer
instead.  It's a relatively new feature of the latest sysklogd release,
where logger got this ability.  The downside is that nothing is output
on the console, the upside is that everything is logged.

For very early script helpers this is a very useful feature, because
when sysklogd eventually starts up it empties the kernel ring buffer and
notices that these messages are not from the kernel, and redirects them
to the proper log file in /var/log.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>