package/finit: disable built-in watchdog daemon by default
In NetBox we use watchdogd for all OS builds, and the app builds don't
need a watchdog.  If they do there's always the simple one that's in
BusyBox.

This prevents issues when watchdogd is built before Finit header files
are installed in staging and the watchdogd configure script cannot by
itself detect a Finit capable system -- which will lead to watchdogd not
being able to start properly since it doesn't know how to tell Finit to
please disable its watchdog service.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>