coronet: mdio: Disable preamble generation, set MDC to 2.5 MHz
No device on the bus requires any preamble to be generated, so disable
it.

Set the MDC clock to 2.5 MHz, speeds above 4 MHz seem to trigger
erratum A-009885.

NOTE: The upstream driver does not support this yet, but the DT
bindings for these settings are already standardized, so there should
be no integration issues once the features are added.