buildroot: bump to 2022.02-dev (LTS), g35a3c01824
This brings in the long awaited iproute2 v5.16.0 to allow for setting
bridge global VLAN and per-VLAN settings introduced in Linux 5.15. In
particular this means we can now configure IGMP snooping per VLAN as
planned for upcoming internal releases.

There are quite a few other changes as well, but should not affect the
overall stability of NetBox.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>