Enable RSTP by default on br0
When mstpd is enabled in the platform's configuration we enable br0 in
/etc/bridge-stp.conf and start mstpd at boot.

The default stp_state is now set to 1 (enabled), meaning if mstpd is not
enabled in the configuration we fall back to STP (default in bridge), if
mstpd is available it will tell the kernel to enable user_stp by setting
stp_state to zero, which causes the kernel (!) to call /sbin/bridge-stp.

All bridge ports will by default be set in the auto-edge mode, giving us
swift forwarding time on access ports, without risking storm.

Notice the current solution with a 'sysv' process in the Finit script,
this means we cannot monitor the mstpd service.  Proper integartion will
require reworking this, but remember the kernel mech. above.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>