board/app: enable tools and fstab for app builds
This patch complements the previous ones in this set, bringing it all
together so that app image builds will work just like their cousins.

 - Enable Busybox iproute2 compat tools: ip addr, ip route, ip link
   etc. which are used by system bringup scripts in board/common
 - A separate /etc/fstab, we inherit a lot from an LXC host so no need
   to double-mount or rely on some thing that are not ours to know
 - A separate /etc/network/interfaces that has DHCP enabled by default

I have blogged a bit about how to run a Zero App build from a Linux Mint
host, with the relevant lxc tools.  It's possible to run other archs as
well, using the Linux kernel's binfmt_misc and Qemu, but that's not in
the blog:

  https://troglobit.com/post/2022-02-13-finit-in-linux-containers-lxc/

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>