patches/busybox: add missing '-i' flag to BusyBox logger tool
The BusyBox logger tool was missing the standard '-i' command line
switch, which adds PID to syslog messages.  Not just the sysklogd
project's logger, but also the standard util-linux logger support
this command line switch.

This patch adds the -i switch and thus fixes the issue with missing
logs from the udhcpc.script in NetBox.

This patch should be upstreamed to the BusyBox project.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>