board/common: adapt mdev-like-a-boss.conf to Buildroot groups
The Buildroot default groups are derived from Debian/Ubuntu, whereas the
mdev-like-a-boss groups used in their mdev.conf seem to be from Fedora.

This patch aligns mdev.conf with Buildroot default group names and also
disables a few kvm related rules.  The idea is to be able to implement
mulitple users as easily as possible using group membership instead of
having all users be root.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>