[BUG REPORT]: add example to recreate [See details]
Building the package `modules` (`make modules-reconfigure`) causes Make to
segfault with the binary shipped with Ubuntu 22.04, but not the one
shipped with Ubuntu 21.10. It also works fine on Ubuntu 20.04 with
Make-4.2.1.

I had a colleague send me his binary from
21.10 and it works fine, while mine on 22.04 does not. Both are
Make-4.3. Building Make-4.3 from source also causes the segfault.
Building latest Make from source does not cause any issue.

The cause of the segfault is at `src/function.c` at function
`func_filter_filterout` at the following code. The alloca keeps
allocating until it goes outside memory and causes write outside its own
stack.

while ((p = find_next_token (&word_iterator, &len)) != 0) {
	struct a_word *word = alloca (sizeof (struct a_word));

	*wordtail = word;
	wordtail = &word->next;

	if (*word_iterator != '\0')
		++word_iterator;

	p[len] = '\0';
	word->str = p;
	word->length = len;
	word->matched = 0;
	word->chain = 0;
	words++;
}

Though, the root cause seems to be different because this code is the
same in the different versions of Ubuntu's Make source code. When it
segfaults it seems to be iterating over many things which it should not
iterate over.

Regarding Buildroot, it seems to be failing between the Configuration
and the Build step. It finishes Configuration, but never reaches the
start of Build. This only seems to happen when `modules` is pointing to
a separate folder of modules. If I put the `hello` module as a normal
package there is no issue building it.

Run the commands below to recreate. Note that it will take some time, as
it has to build the Linux kernel.
$ git clone https://github.com/westermo/netbox/tree/make-bug
$ cd netbox
$ make netbox_os_zero_defconfig
$ make
$ make modules-reconfigure # segfault happens here

After that, the last command can be repeated to test building the
modules.